+++ card-game-web/src/components/player/PrimaryPlayer.jsx
@@ -2,11 +2,11 @@ import React from "react";
 import { useSelector } from "react-redux";
 import { selectIsSpectating } from "../../store/lobby/lobby-store";
 import { ORIENTATIONS } from "../card/CardStump";
+import Deck from "../card/Deck";
+import DiscardPile from "../card/DiscardPile";
 import * as c from "../styles/colors";
 import styled from "styled-components";
 import * as S from "../../styles";
-import Deck from "../card/Deck";
-import Pile from "../card/Pile";
 import Structures from "../structure/Structures";
 import Hand from "./Hand";
 import Hero from "./Hero";
@@ -46,14 +46,14 @@ const PrimaryPlayer = ({ player, active }) => {
       </SpanAll>
       <Spacer />
       <SubGrid>
-        <Deck size={player.deckSize} />
+        <Deck cards={player.deck} />
         <Hero
           entityId={player.entityId}
           name={player.user}
           health={player.health}
           active={active}
         />
-        <Pile cards={player.discardPile} name="discard pile" />
+        <DiscardPile cards={player.discardPile} />
       </SubGrid>
       <Spacer />
     </>